# openapi.yaml
...
openapi: 3.0.0
...
info:
  title: CRUD Master API Gateway
  description: |-
    Complete API Gateway for movie inventory and billing microservices.

    This API provides:
    - **Movie Management**: Full CRUD operations for movie inventory
    - **Order Processing**: Asynchronous order creation via RabbitMQ

    ## Architecture
    - API Gateway (Port 3000) serves as the single entry point
    - HTTP proxy routes inventory requests to Inventory API (Port 8080)
    - Billing requests are published to RabbitMQ queue (async processing)

    ## Authentication
    Currently no authentication required for testing. In production, add JWT/OAuth2.

  version: 1.0.0
  contact:
    name: Saddam Hussain
    email: saddam.hussain@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

# ============================================================================================
# SERVERS BLOCK - Where the API Can Be Accessed
# ============================================================================================
# Purpose: Lists all available API endpoints (development, staging, production)
# Used by: API clients to know where to send requests
servers:
  - url: http://192.168.56.10:3000
    description: API Gateway (production - Vagrant VM)
  - url: http://localhost:3000
    description: API Gateway (local development)

# ============================================================================================
...
paths:
...
  /api/movies:
    # =======================================================================================
    # HTTP GET /api/movies - Retrieve List of All Movies
    # =======================================================================================
    # Purpose: Fetch all movies from inventory, with optional filtering by title
    # HTTP Method: GET (read-only, safe operation)
    # Response: 200 OK with array of movie objects, or 500 on server error
    get:
      summary: Get all movies
      description: |-
        Retrieve a list of all movies in the inventory.

        Supports optional filtering by movie title.
      operationId: getAllMovies
      tags:
        - Movies
      parameters:
        - name: title
          in: query
          description: Filter movies by title (case-insensitive, partial match)
          required: false
          schema:
            type: string
            example: Inception
      responses:
        "200":
          description: Successfully retrieved movies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieListResponse"
              example:
                success: true
                count: 2
                movies:
                  - id: 1
                    title: Inception
                    description: A mind-bending thriller
                    genre: Sci-Fi
                    release_year: 2010
                    rating: 8.8
                    duration: 148
                    available_copies: 5
                    created_at: "2026-03-27T18:01:12.000000"
                    updated_at: "2026-03-27T18:01:12.000000"
                  - id: 2
                    title: Titanic
                    description: A love story on a sinking ship
                    genre: Romance
                    release_year: 1997
                    rating: 7.8
                    duration: 194
                    available_copies: 4
                    created_at: "2026-03-27T18:02:00.000000"
                    updated_at: "2026-03-27T18:02:00.000000"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    # =======================================================================================
    # HTTP POST /api/movies - Create a New Movie
    # =======================================================================================
    # Purpose: Add a new movie to the inventory database
    # HTTP Method: POST (creates new resource)
    # Response: 201 Created with the new movie object, or 400/500 on error
    post:
      summary: Create a new movie
      description: |-
        Create a new movie in the inventory.

        Required fields: title, genre, release_year

        Returns the created movie with auto-generated ID and timestamps.
      operationId: createMovie
      tags:
        - Movies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMovieRequest"
            example:
              title: Inception
              description: A mind-bending thriller about dreams
              genre: Sci-Fi
              release_year: 2010
              rating: 8.8
              duration: 148
              available_copies: 5
      responses:
        "201":
          description: Movie created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieResponse"
              example:
                success: true
                message: Movie created successfully
                movie:
                  id: 1
                  title: Inception
                  description: A mind-bending thriller
                  genre: Sci-Fi
                  release_year: 2010
                  rating: 8.8
                  duration: 148
                  available_copies: 5
                  created_at: "2026-03-27T18:01:12.000000"
                  updated_at: "2026-03-27T18:01:12.000000"
        "400":
          description: Invalid input (missing required fields or invalid data type)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    # =======================================================================================
    # HTTP DELETE /api/movies - Delete ALL Movies (Destructive Operation)
    # =======================================================================================
    # Purpose: Remove ALL movies from the database in a single operation
    # HTTP Method: DELETE (destructive, cannot be undone)
    # ⚠️ WARNING: This is a bulk operation - use with extreme caution!
    # Response: 200 OK with count of deleted movies, or 500 on error
    delete:
      summary: Delete all movies
      description: |-
        Delete ALL movies from the database.

        ⚠️ **WARNING**: This is a destructive operation and cannot be undone!

        Returns the count of deleted movies.
      operationId: deleteAllMovies
      tags:
        - Movies
      responses:
        "200":
          description: All movies deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteAllMoviesResponse"
              example:
                success: true
                message: All movies deleted successfully
                deleted_count: 5
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ==========================================================================================
  # ENDPOINT: /api/movies/{id}
  # ==========================================================================================
  # Purpose: Operations on a SPECIFIC movie (use ID in URL path)
  # Supports: GET (fetch), PUT (update), DELETE (remove)
  # Use Case: Get details, update fields, or delete one movie
  /api/movies/{id}:
    # =======================================================================================
    # HTTP GET /api/movies/{id} - Retrieve a Single Movie by ID
    # =======================================================================================
    # Purpose: Fetch details of one specific movie using its unique ID
    # HTTP Method: GET (read-only)
    # URL Parameter: id (integer) - The movie's unique identifier
    # Response: 200 OK with movie object, or 404 if movie doesn't exist
    get:
      summary: Get a single movie by ID
      description: Retrieve a specific movie by its unique ID.
      operationId: getMovieById
      tags:
        - Movies
      parameters:
        - name: id
          in: path
          description: Unique movie ID
          required: true
          schema:
            type: integer
            example: 1
      responses:
        "200":
          description: Movie found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieResponse"
              example:
                success: true
                movie:
                  id: 1
                  title: Inception
                  description: A mind-bending thriller
                  genre: Sci-Fi
                  release_year: 2010
                  rating: 8.8
                  duration: 148
                  available_copies: 5
                  created_at: "2026-03-27T18:01:12.000000"
                  updated_at: "2026-03-27T18:01:12.000000"
        "404":
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                error: "Movie with ID 999 not found"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    # =======================================================================================
    # HTTP PUT /api/movies/{id} - Update a Specific Movie
    # =======================================================================================
    # Purpose: Modify one or more fields of an existing movie
    # HTTP Method: PUT (updates existing resource)
    # URL Parameter: id (integer) - The movie to update
    # Response: 200 OK with updated movie object, or 404/400/500 on error
    # Special Feature: Supports partial updates (only send fields you want to change)
    put:
      summary: Update a movie
      description: |-
        Update a specific movie by ID.

        Only provided fields will be updated (partial updates supported).

        Timestamps (created_at, updated_at) are automatically managed.
      operationId: updateMovie
      tags:
        - Movies
      parameters:
        - name: id
          in: path
          description: Unique movie ID
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMovieRequest"
            example:
              rating: 9.0
              available_copies: 3
      responses:
        "200":
          description: Movie updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieResponse"
              example:
                success: true
                message: Movie updated successfully
                movie:
                  id: 1
                  title: Inception
                  description: A mind-bending thriller
                  genre: Sci-Fi
                  release_year: 2010
                  rating: 9.0
                  duration: 148
                  available_copies: 3
                  created_at: "2026-03-27T18:01:12.000000"
                  updated_at: "2026-03-27T18:15:45.000000"
        "400":
          description: Invalid input (invalid data type)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Movie not found
          content:
            application/json:
              schema:
                # =======================================================================================
                # HTTP DELETE /api/movies/{id} - Delete a Single Movie
                # =======================================================================================
                # Purpose: Remove one specific movie from the database
                # HTTP Method: DELETE (destructive)
                # URL Parameter: id (integer) - The movie to delete
                # Response: 200 OK if deleted, or 404 if movie doesn't exist
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      summary: Delete a specific movie
      description: Delete a movie by its unique ID.
      operationId: deleteMovie
      tags:
        - Movies
      parameters:
        - name: id
          in: path
          description: Unique movie ID
          required: true
          schema:
            type: integer
            example: 1
      responses:
        "200":
          description: Movie deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Movie deleted successfully
        "404":
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ==========================================================================================
  # ENDPOINT: /api/billing
  # ==========================================================================================
  # Purpose: Asynchronous order creation endpoint (publishes to RabbitMQ queue)
  # Pattern: Message queue (fire-and-forget), NOT HTTP proxy
  # Key Feature: Returns 200 immediately, processing happens async by Billing API
  # Resilience: Orders persist in queue even if Billing API is offline
...
components:
  schemas:
    Movie:
      type: object
      required:
        - id
        - title
        - genre
        - release_year
      properties:
        id:
          type: integer
          description: Unique movie ID (auto-generated)
          example: 1
        title:
          type: string
          description: Movie title
          example: Inception
        description:
          type: string
          nullable: true
          description: Movie synopsis (optional)
          example: A mind-bending thriller about dreams
        genre:
          type: string
          description: Movie genre
          example: Sci-Fi
        release_year:
          type: integer
          description: Year the movie was released
          example: 2010
        rating:
          type: number
          format: float
          description: Movie rating (0-10, optional)
          example: 8.8
        duration:
          type: integer
          description: Movie duration in minutes
          example: 148
        available_copies:
          type: integer
          description: Number of available copies
          example: 5
        created_at:
          type: string
          format: date-time
          description: ISO 8601 timestamp when movie was created
          example: "2026-03-27T18:01:12.000000"
        updated_at:
          type: string
          format: date-time
          description: ISO 8601 timestamp when movie was last updated
          example: "2026-03-27T18:01:12.000000"
    CreateMovieRequest:
      type: object
      required:
        - title
        - genre
        - release_year
      properties:
        title:
          type: string
          description: Movie title (required)
          example: Inception
        description:
          type: string
          description: Movie synopsis (optional)
          example: A mind-bending thriller
        genre:
          type: string
          description: Movie genre (required)
          example: Sci-Fi
        release_year:
          type: integer
          description: Year released (required)
          example: 2010
        rating:
          type: number
          format: float
          description: Rating 0-10 (optional, default 0.0)
          example: 8.8
        duration:
          type: integer
          description: Duration in minutes (required, default 120)
          example: 148
        available_copies:
          type: integer
          description: Number of copies (required, default 0)
          example: 5
    UpdateMovieRequest:
      type: object
      description: Any number of fields can be updated (partial update)
      properties:
        title:
          type: string
          example: Inception
        description:
          type: string
          example: A mind-bending thriller
        genre:
          type: string
          example: Sci-Fi
        release_year:
          type: integer
          example: 2010
        rating:
          type: number
          format: float
          example: 9.0
        duration:
          type: integer
          example: 148
        available_copies:
          type: integer
          example: 3
...
    MovieResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Movie created successfully
        movie:
          $ref: "#/components/schemas/Movie"

    # =======================================================================================
    # SCHEMA: MovieListResponse
    # =======================================================================================
    # Purpose: Data structure for successful multi-movie responses
    # Used By: GET /api/movies (when retrieving all movies)
    # Wrapper: Includes success flag, count, and array of Movie objects
    MovieListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        count:
          type: integer
          example: 2
        movies:
          type: array
          items:
            $ref: "#/components/schemas/Movie"
    DeleteAllMoviesResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: All movies deleted successfully
        deleted_count:
          type: integer
          example: 5
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Movie not found
        message:
          type: string
          example: Additional error details
        details:
          type: string
          example: Additional debug information